requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
deprecation==2.1.0
gotrue==1.0.4
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.1
Pillow==10.0.0
postgrest==0.10.8
pydantic==2.3.0
pydantic_core==2.6.3
pygubu==0.31
python-dateutil==2.8.2
python-dotenv==1.0.0
realtime==1.0.0
six==1.16.0
sniffio==1.3.0
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
typing_extensions==4.7.1
websockets==10.4
